+++ extensions/iceworks-lowcode-page-builder/src/components/Material/components/view.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Tab } from '@alifd/next';
-import { IMaterialSource, IMaterialTypeDatum, IMaterialScaffold, IMaterialBlock, IMaterialComponent, IMaterialBase } from '../types';
+import { IMaterialSource, IMaterialTypeDatum, IMaterialScaffold, IMaterialBlock, IMaterialComponent, IMaterialBase } from '@iceworks/material/lib/common';
 import { MaterialType } from './type';
 import * as styles from './view.module.scss';
