+++ packages/material/src/common/constant.ts
@@ -1,2 +1,2 @@
 export const PLACEHOLDER_IMG = 'https://img.alicdn.com/tfs/TB1lCcKc8Kw3KVjSZFOXXarDVXa-350-200.png';
-export const customCategory = 'iceworks-category';
+export const CUSTOM_CATEGORY = 'iceworks-category';
